# src/tests/contract/reports/HolidayService2-swaggerDoc.yaml
swagger: '2.0'
info:
  title: HolidayService2
  description: Web service that calculates holiday dates. (Version 2.0.1)
  x-ibm-name: holidayservice2
  version: 1.0.0
schemes:
  - https
basePath: /HolidayService2
produces:
  - application/xml
consumes:
  - text/xml
securityDefinitions:
  clientID:
    type: apiKey
    name: X-IBM-Client-Id
    in: header
    description: ''
security:
  - clientID: []
...
paths:
...
  /GetHolidaysAvailable:
    post:
      summary: Operation GetHolidaysAvailable
      description: Get the available holidays for a specified country.
      operationId: GetHolidaysAvailable
      x-ibm-soap:
        soap-action: >-
          http://www.holidaywebservice.com/HolidayService_v2/GetHolidaysAvailable
        soap-operation: >-
          {http://www.holidaywebservice.com/HolidayService_v2/}GetHolidaysAvailable
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/GetHolidaysAvailableInput'
      responses:
        default:
          description: ''
          schema:
            $ref: '#/definitions/GetHolidaysAvailableOutput'
  /GetHolidayDate:
    post:
      summary: Operation GetHolidayDate
      description: Get the date of a specific holiday.
      operationId: GetHolidayDate
      x-ibm-soap:
        soap-action: 'http://www.holidaywebservice.com/HolidayService_v2/GetHolidayDate'
        soap-operation: '{http://www.holidaywebservice.com/HolidayService_v2/}GetHolidayDate'
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/GetHolidayDateInput'
      responses:
        default:
          description: ''
          schema:
            $ref: '#/definitions/GetHolidayDateOutput'
  /GetHolidaysForDateRange:
    post:
      summary: Operation GetHolidaysForDateRange
      description: Get the holidays for a date range.
      operationId: GetHolidaysForDateRange
      x-ibm-soap:
        soap-action: >-
          http://www.holidaywebservice.com/HolidayService_v2/GetHolidaysForDateRange
        soap-operation: >-
          {http://www.holidaywebservice.com/HolidayService_v2/}GetHolidaysForDateRange
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/GetHolidaysForDateRangeInput'
      responses:
        default:
          description: ''
          schema:
            $ref: '#/definitions/GetHolidaysForDateRangeOutput'
  /GetHolidaysForYear:
    post:
      summary: Operation GetHolidaysForYear
      description: Get the holidays for an entire year.
      operationId: GetHolidaysForYear
      x-ibm-soap:
        soap-action: 'http://www.holidaywebservice.com/HolidayService_v2/GetHolidaysForYear'
        soap-operation: >-
          {http://www.holidaywebservice.com/HolidayService_v2/}GetHolidaysForYear
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/GetHolidaysForYearInput'
      responses:
        default:
          description: ''
          schema:
            $ref: '#/definitions/GetHolidaysForYearOutput'
  /GetHolidaysForMonth:
    post:
      summary: Operation GetHolidaysForMonth
      description: Get the holidays for a specific month.
      operationId: GetHolidaysForMonth
      x-ibm-soap:
        soap-action: 'http://www.holidaywebservice.com/HolidayService_v2/GetHolidaysForMonth'
        soap-operation: >-
          {http://www.holidaywebservice.com/HolidayService_v2/}GetHolidaysForMonth
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/GetHolidaysForMonthInput'
      responses:
        default:
          description: ''
          schema:
            $ref: '#/definitions/GetHolidaysForMonthOutput'
definitions:
...
  GetHolidaysAvailableInput:
    type: object
    properties:
      Envelope:
        xml:
          namespace: 'http://schemas.xmlsoap.org/soap/envelope/'
          prefix: soapenv
        type: object
        properties:
          Header:
            $ref: '#/definitions/GetHolidaysAvailableHeader'
          Body:
            xml:
              namespace: 'http://schemas.xmlsoap.org/soap/envelope/'
              prefix: soapenv
            type: object
            properties:
              GetHolidaysAvailable:
                $ref: '#/definitions/GetHolidaysAvailable_tns'
            required:
              - GetHolidaysAvailable
        required:
          - Body
    required:
      - Envelope
    x-ibm-schema:
      wsdl-port: '{http://www.holidaywebservice.com/HolidayService_v2/}HolidayService2Soap'
      wsd-operation: GetHolidaysAvailable
      wsdl-message-direction-or-name: GetHolidaysAvailableRequest
    example: >-

      <soapenv:Envelope
      xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/">
       <soapenv:Header>
        <wsse:Security xmlns:wsse="http://docs.oasis-open.org/wss/2004/01/oasis-200401-wss-wssecurity-secext-1.0.xsd" xmlns:wsu="http://docs.oasis-open.org/wss/2004/01/oasis-200401-wss-wssecurity-utility-1.0.xsd">
         <wsse:UsernameToken>
          <wsse:Username>string</wsse:Username>
          <wsse:Password>string</wsse:Password>
          <wsse:Nonce EncodingType="string">string</wsse:Nonce>
          <wsu:Created>string</wsu:Created>
         </wsse:UsernameToken>
         <wsu:Timestamp wsu:Id="string">
          <wsu:Created>string</wsu:Created>
          <wsu:Expires>string</wsu:Expires>
         </wsu:Timestamp>
        </wsse:Security>
       </soapenv:Header>
       <soapenv:Body>
        <tns:GetHolidaysAvailable xmlns:tns="http://www.holidaywebservice.com/HolidayService_v2/"><!-- mandatory -->
         <tns:countryCode><!-- mandatory -->string</tns:countryCode>
        </tns:GetHolidaysAvailable>
       </soapenv:Body>
      </soapenv:Envelope>
  GetHolidaysAvailableHeader:
    type: object
    properties:
      Security:
        $ref: '#/definitions/Security'
  GetHolidaysAvailableOutput:
    type: object
    properties:
      Envelope:
        xml:
          namespace: 'http://schemas.xmlsoap.org/soap/envelope/'
          prefix: soapenv
        type: object
        properties:
          Body:
            xml:
              namespace: 'http://schemas.xmlsoap.org/soap/envelope/'
              prefix: soapenv
            type: object
            properties:
              GetHolidaysAvailableResponse:
                $ref: '#/definitions/GetHolidaysAvailableResponse_tns'
        required:
          - Body
    required:
      - Envelope
    x-ibm-schema:
      wsdl-port: '{http://www.holidaywebservice.com/HolidayService_v2/}HolidayService2Soap'
      wsd-operation: GetHolidaysAvailable
      wsdl-message-direction-or-name: GetHolidaysAvailableResponse
    example: >-

      <soapenv:Envelope
      xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/">
       <soapenv:Body>
        <tns:GetHolidaysAvailableResponse xmlns:tns="http://www.holidaywebservice.com/HolidayService_v2/">
         <tns:GetHolidaysAvailableResult>
          <tns:HolidayCode><!-- between zero and unlimited repetitions of this element -->
           <tns:Code>string</tns:Code>
           <tns:Description>string</tns:Description>
          </tns:HolidayCode>
         </tns:GetHolidaysAvailableResult>
        </tns:GetHolidaysAvailableResponse>
       </soapenv:Body>
      </soapenv:Envelope>
  GetHolidayDateInput:
    type: object
    properties:
      Envelope:
        xml:
          namespace: 'http://schemas.xmlsoap.org/soap/envelope/'
          prefix: soapenv
        type: object
        properties:
          Header:
            $ref: '#/definitions/GetHolidayDateHeader'
          Body:
            xml:
              namespace: 'http://schemas.xmlsoap.org/soap/envelope/'
              prefix: soapenv
            type: object
            properties:
              GetHolidayDate:
                $ref: '#/definitions/GetHolidayDate_tns'
            required:
              - GetHolidayDate
        required:
          - Body
    required:
      - Envelope
    x-ibm-schema:
      wsdl-port: '{http://www.holidaywebservice.com/HolidayService_v2/}HolidayService2Soap'
      wsd-operation: GetHolidayDate
      wsdl-message-direction-or-name: GetHolidayDateRequest
    example: >-

      <soapenv:Envelope
      xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/">
       <soapenv:Header>
        <wsse:Security xmlns:wsse="http://docs.oasis-open.org/wss/2004/01/oasis-200401-wss-wssecurity-secext-1.0.xsd" xmlns:wsu="http://docs.oasis-open.org/wss/2004/01/oasis-200401-wss-wssecurity-utility-1.0.xsd">
         <wsse:UsernameToken>
          <wsse:Username>string</wsse:Username>
          <wsse:Password>string</wsse:Password>
          <wsse:Nonce EncodingType="string">string</wsse:Nonce>
          <wsu:Created>string</wsu:Created>
         </wsse:UsernameToken>
         <wsu:Timestamp wsu:Id="string">
          <wsu:Created>string</wsu:Created>
          <wsu:Expires>string</wsu:Expires>
         </wsu:Timestamp>
        </wsse:Security>
       </soapenv:Header>
       <soapenv:Body>
        <tns:GetHolidayDate xmlns:tns="http://www.holidaywebservice.com/HolidayService_v2/"><!-- mandatory -->
         <tns:countryCode><!-- mandatory -->string</tns:countryCode>
         <tns:holidayCode>string</tns:holidayCode>
         <tns:year><!-- mandatory -->3</tns:year>
        </tns:GetHolidayDate>
       </soapenv:Body>
      </soapenv:Envelope>
  GetHolidayDateHeader:
    type: object
    properties:
      Security:
        $ref: '#/definitions/Security'
  GetHolidayDateOutput:
    type: object
    properties:
      Envelope:
        xml:
          namespace: 'http://schemas.xmlsoap.org/soap/envelope/'
          prefix: soapenv
        type: object
        properties:
          Body:
            xml:
              namespace: 'http://schemas.xmlsoap.org/soap/envelope/'
              prefix: soapenv
            type: object
            properties:
              GetHolidayDateResponse:
                $ref: '#/definitions/GetHolidayDateResponse_tns'
        required:
          - Body
    required:
      - Envelope
    x-ibm-schema:
      wsdl-port: '{http://www.holidaywebservice.com/HolidayService_v2/}HolidayService2Soap'
      wsd-operation: GetHolidayDate
      wsdl-message-direction-or-name: GetHolidayDateResponse
    example: >-

      <soapenv:Envelope
      xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/">
       <soapenv:Body>
        <tns:GetHolidayDateResponse xmlns:tns="http://www.holidaywebservice.com/HolidayService_v2/">
         <tns:GetHolidayDateResult><!-- mandatory -->2016-04-18T14:07:37</tns:GetHolidayDateResult>
        </tns:GetHolidayDateResponse>
       </soapenv:Body>
      </soapenv:Envelope>
  GetHolidaysForDateRangeInput:
    type: object
    properties:
      Envelope:
        xml:
          namespace: 'http://schemas.xmlsoap.org/soap/envelope/'
          prefix: soapenv
        type: object
        properties:
          Header:
            $ref: '#/definitions/GetHolidaysForDateRangeHeader'
          Body:
            xml:
              namespace: 'http://schemas.xmlsoap.org/soap/envelope/'
              prefix: soapenv
            type: object
            properties:
              GetHolidaysForDateRange:
                $ref: '#/definitions/GetHolidaysForDateRange_tns'
            required:
              - GetHolidaysForDateRange
        required:
          - Body
    required:
      - Envelope
    x-ibm-schema:
      wsdl-port: '{http://www.holidaywebservice.com/HolidayService_v2/}HolidayService2Soap'
      wsd-operation: GetHolidaysForDateRange
      wsdl-message-direction-or-name: GetHolidaysForDateRangeRequest
    example: >-

      <soapenv:Envelope
      xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/">
       <soapenv:Header>
        <wsse:Security xmlns:wsse="http://docs.oasis-open.org/wss/2004/01/oasis-200401-wss-wssecurity-secext-1.0.xsd" xmlns:wsu="http://docs.oasis-open.org/wss/2004/01/oasis-200401-wss-wssecurity-utility-1.0.xsd">
         <wsse:UsernameToken>
          <wsse:Username>string</wsse:Username>
          <wsse:Password>string</wsse:Password>
          <wsse:Nonce EncodingType="string">string</wsse:Nonce>
          <wsu:Created>string</wsu:Created>
         </wsse:UsernameToken>
         <wsu:Timestamp wsu:Id="string">
          <wsu:Created>string</wsu:Created>
          <wsu:Expires>string</wsu:Expires>
         </wsu:Timestamp>
        </wsse:Security>
       </soapenv:Header>
       <soapenv:Body>
        <tns:GetHolidaysForDateRange xmlns:tns="http://www.holidaywebservice.com/HolidayService_v2/"><!-- mandatory -->
         <tns:countryCode><!-- mandatory -->string</tns:countryCode>
         <tns:startDate><!-- mandatory -->2016-04-18T14:07:37</tns:startDate>
         <tns:endDate><!-- mandatory -->2016-04-18T14:07:37</tns:endDate>
        </tns:GetHolidaysForDateRange>
       </soapenv:Body>
      </soapenv:Envelope>
  GetHolidaysForDateRangeHeader:
    type: object
    properties:
      Security:
        $ref: '#/definitions/Security'
  GetHolidaysForDateRangeOutput:
    type: object
    properties:
      Envelope:
        xml:
          namespace: 'http://schemas.xmlsoap.org/soap/envelope/'
          prefix: soapenv
        type: object
        properties:
          Body:
            xml:
              namespace: 'http://schemas.xmlsoap.org/soap/envelope/'
              prefix: soapenv
            type: object
            properties:
              GetHolidaysForDateRangeResponse:
                $ref: '#/definitions/GetHolidaysForDateRangeResponse_tns'
        required:
          - Body
    required:
      - Envelope
    x-ibm-schema:
      wsdl-port: '{http://www.holidaywebservice.com/HolidayService_v2/}HolidayService2Soap'
      wsd-operation: GetHolidaysForDateRange
      wsdl-message-direction-or-name: GetHolidaysForDateRangeResponse
    example: >-

      <soapenv:Envelope
      xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/">
       <soapenv:Body>
        <tns:GetHolidaysForDateRangeResponse xmlns:tns="http://www.holidaywebservice.com/HolidayService_v2/">
         <tns:GetHolidaysForDateRangeResult>
          <tns:Holiday><!-- between zero and unlimited repetitions of this element -->
           <tns:Country><!-- mandatory -->string</tns:Country>
           <tns:HolidayCode>string</tns:HolidayCode>
           <tns:Descriptor>string</tns:Descriptor>
           <tns:HolidayType><!-- mandatory -->string</tns:HolidayType>
           <tns:DateType><!-- mandatory -->string</tns:DateType>
           <tns:BankHoliday><!-- mandatory -->string</tns:BankHoliday>
           <tns:Date><!-- mandatory -->2016-04-18T14:07:37</tns:Date>
           <tns:RelatedHolidayCode>string</tns:RelatedHolidayCode>
           <tns:ApplicableRegions>
            <tns:RegionCode><!-- between zero and unlimited repetitions of this element -->
             <tns:Code>string</tns:Code>
             <tns:Description>string</tns:Description>
            </tns:RegionCode>
           </tns:ApplicableRegions>
          </tns:Holiday>
         </tns:GetHolidaysForDateRangeResult>
        </tns:GetHolidaysForDateRangeResponse>
       </soapenv:Body>
      </soapenv:Envelope>
  GetHolidaysForYearInput:
    type: object
    properties:
      Envelope:
        xml:
          namespace: 'http://schemas.xmlsoap.org/soap/envelope/'
          prefix: soapenv
        type: object
        properties:
          Header:
            $ref: '#/definitions/GetHolidaysForYearHeader'
          Body:
            xml:
              namespace: 'http://schemas.xmlsoap.org/soap/envelope/'
              prefix: soapenv
            type: object
            properties:
              GetHolidaysForYear:
                $ref: '#/definitions/GetHolidaysForYear_tns'
            required:
              - GetHolidaysForYear
        required:
          - Body
    required:
      - Envelope
    x-ibm-schema:
      wsdl-port: '{http://www.holidaywebservice.com/HolidayService_v2/}HolidayService2Soap'
      wsd-operation: GetHolidaysForYear
      wsdl-message-direction-or-name: GetHolidaysForYearRequest
    example: >-

      <soapenv:Envelope
      xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/">
       <soapenv:Header>
        <wsse:Security xmlns:wsse="http://docs.oasis-open.org/wss/2004/01/oasis-200401-wss-wssecurity-secext-1.0.xsd" xmlns:wsu="http://docs.oasis-open.org/wss/2004/01/oasis-200401-wss-wssecurity-utility-1.0.xsd">
         <wsse:UsernameToken>
          <wsse:Username>string</wsse:Username>
          <wsse:Password>string</wsse:Password>
          <wsse:Nonce EncodingType="string">string</wsse:Nonce>
          <wsu:Created>string</wsu:Created>
         </wsse:UsernameToken>
         <wsu:Timestamp wsu:Id="string">
          <wsu:Created>string</wsu:Created>
          <wsu:Expires>string</wsu:Expires>
         </wsu:Timestamp>
        </wsse:Security>
       </soapenv:Header>
       <soapenv:Body>
        <tns:GetHolidaysForYear xmlns:tns="http://www.holidaywebservice.com/HolidayService_v2/"><!-- mandatory -->
         <tns:countryCode><!-- mandatory -->string</tns:countryCode>
         <tns:year><!-- mandatory -->3</tns:year>
        </tns:GetHolidaysForYear>
       </soapenv:Body>
      </soapenv:Envelope>
  GetHolidaysForYearHeader:
    type: object
    properties:
      Security:
        $ref: '#/definitions/Security'
  GetHolidaysForYearOutput:
    type: object
    properties:
      Envelope:
        xml:
          namespace: 'http://schemas.xmlsoap.org/soap/envelope/'
          prefix: soapenv
        type: object
        properties:
          Body:
            xml:
              namespace: 'http://schemas.xmlsoap.org/soap/envelope/'
              prefix: soapenv
            type: object
            properties:
              GetHolidaysForYearResponse:
                $ref: '#/definitions/GetHolidaysForYearResponse_tns'
        required:
          - Body
    required:
      - Envelope
    x-ibm-schema:
      wsdl-port: '{http://www.holidaywebservice.com/HolidayService_v2/}HolidayService2Soap'
      wsd-operation: GetHolidaysForYear
      wsdl-message-direction-or-name: GetHolidaysForYearResponse
    example: >-

      <soapenv:Envelope
      xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/">
       <soapenv:Body>
        <tns:GetHolidaysForYearResponse xmlns:tns="http://www.holidaywebservice.com/HolidayService_v2/">
         <tns:GetHolidaysForYearResult>
          <tns:Holiday><!-- between zero and unlimited repetitions of this element -->
           <tns:Country><!-- mandatory -->string</tns:Country>
           <tns:HolidayCode>string</tns:HolidayCode>
           <tns:Descriptor>string</tns:Descriptor>
           <tns:HolidayType><!-- mandatory -->string</tns:HolidayType>
           <tns:DateType><!-- mandatory -->string</tns:DateType>
           <tns:BankHoliday><!-- mandatory -->string</tns:BankHoliday>
           <tns:Date><!-- mandatory -->2016-04-18T14:07:37</tns:Date>
           <tns:RelatedHolidayCode>string</tns:RelatedHolidayCode>
           <tns:ApplicableRegions>
            <tns:RegionCode><!-- between zero and unlimited repetitions of this element -->
             <tns:Code>string</tns:Code>
             <tns:Description>string</tns:Description>
            </tns:RegionCode>
           </tns:ApplicableRegions>
          </tns:Holiday>
         </tns:GetHolidaysForYearResult>
        </tns:GetHolidaysForYearResponse>
       </soapenv:Body>
      </soapenv:Envelope>
  GetHolidaysForMonthInput:
    type: object
    properties:
      Envelope:
        xml:
          namespace: 'http://schemas.xmlsoap.org/soap/envelope/'
          prefix: soapenv
        type: object
        properties:
          Header:
            $ref: '#/definitions/GetHolidaysForMonthHeader'
          Body:
            xml:
              namespace: 'http://schemas.xmlsoap.org/soap/envelope/'
              prefix: soapenv
            type: object
            properties:
              GetHolidaysForMonth:
                $ref: '#/definitions/GetHolidaysForMonth_tns'
            required:
              - GetHolidaysForMonth
        required:
          - Body
    required:
      - Envelope
    x-ibm-schema:
      wsdl-port: '{http://www.holidaywebservice.com/HolidayService_v2/}HolidayService2Soap'
      wsd-operation: GetHolidaysForMonth
      wsdl-message-direction-or-name: GetHolidaysForMonthRequest
    example: >-

      <soapenv:Envelope
      xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/">
       <soapenv:Header>
        <wsse:Security xmlns:wsse="http://docs.oasis-open.org/wss/2004/01/oasis-200401-wss-wssecurity-secext-1.0.xsd" xmlns:wsu="http://docs.oasis-open.org/wss/2004/01/oasis-200401-wss-wssecurity-utility-1.0.xsd">
         <wsse:UsernameToken>
          <wsse:Username>string</wsse:Username>
          <wsse:Password>string</wsse:Password>
          <wsse:Nonce EncodingType="string">string</wsse:Nonce>
          <wsu:Created>string</wsu:Created>
         </wsse:UsernameToken>
         <wsu:Timestamp wsu:Id="string">
          <wsu:Created>string</wsu:Created>
          <wsu:Expires>string</wsu:Expires>
         </wsu:Timestamp>
        </wsse:Security>
       </soapenv:Header>
       <soapenv:Body>
        <tns:GetHolidaysForMonth xmlns:tns="http://www.holidaywebservice.com/HolidayService_v2/"><!-- mandatory -->
         <tns:countryCode><!-- mandatory -->string</tns:countryCode>
         <tns:year><!-- mandatory -->3</tns:year>
         <tns:month><!-- mandatory -->3</tns:month>
        </tns:GetHolidaysForMonth>
       </soapenv:Body>
      </soapenv:Envelope>
  GetHolidaysForMonthHeader:
    type: object
    properties:
      Security:
        $ref: '#/definitions/Security'
  GetHolidaysForMonthOutput:
    type: object
    properties:
      Envelope:
        xml:
          namespace: 'http://schemas.xmlsoap.org/soap/envelope/'
          prefix: soapenv
        type: object
        properties:
          Body:
            xml:
              namespace: 'http://schemas.xmlsoap.org/soap/envelope/'
              prefix: soapenv
            type: object
            properties:
              GetHolidaysForMonthResponse:
                $ref: '#/definitions/GetHolidaysForMonthResponse_tns'
        required:
          - Body
    required:
      - Envelope
    x-ibm-schema:
      wsdl-port: '{http://www.holidaywebservice.com/HolidayService_v2/}HolidayService2Soap'
      wsd-operation: GetHolidaysForMonth
      wsdl-message-direction-or-name: GetHolidaysForMonthResponse
    example: >-

      <soapenv:Envelope
      xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/">
       <soapenv:Body>
        <tns:GetHolidaysForMonthResponse xmlns:tns="http://www.holidaywebservice.com/HolidayService_v2/">
         <tns:GetHolidaysForMonthResult>
          <tns:Holiday><!-- between zero and unlimited repetitions of this element -->
           <tns:Country><!-- mandatory -->string</tns:Country>
           <tns:HolidayCode>string</tns:HolidayCode>
           <tns:Descriptor>string</tns:Descriptor>
           <tns:HolidayType><!-- mandatory -->string</tns:HolidayType>
           <tns:DateType><!-- mandatory -->string</tns:DateType>
           <tns:BankHoliday><!-- mandatory -->string</tns:BankHoliday>
           <tns:Date><!-- mandatory -->2016-04-18T14:07:37</tns:Date>
           <tns:RelatedHolidayCode>string</tns:RelatedHolidayCode>
           <tns:ApplicableRegions>
            <tns:RegionCode><!-- between zero and unlimited repetitions of this element -->
             <tns:Code>string</tns:Code>
             <tns:Description>string</tns:Description>
            </tns:RegionCode>
           </tns:ApplicableRegions>
          </tns:Holiday>
         </tns:GetHolidaysForMonthResult>
        </tns:GetHolidaysForMonthResponse>
       </soapenv:Body>
      </soapenv:Envelope>
...
  GetHolidaysAvailable_tns:
    xml:
      namespace: 'http://www.holidaywebservice.com/HolidayService_v2/'
      prefix: tns
    type: object
    properties:
      countryCode:
        $ref: '#/definitions/Country_tns'
    required:
      - countryCode
    example: >-

      <tns:GetHolidaysAvailable
      xmlns:tns="http://www.holidaywebservice.com/HolidayService_v2/">
       <tns:countryCode><!-- mandatory -->string</tns:countryCode>
      </tns:GetHolidaysAvailable>
  GetHolidaysAvailableResponse_tns:
    xml:
      namespace: 'http://www.holidaywebservice.com/HolidayService_v2/'
      prefix: tns
    type: object
    properties:
      GetHolidaysAvailableResult:
        $ref: '#/definitions/ArrayOfHolidayCode_tns'
    example: >-

      <tns:GetHolidaysAvailableResponse
      xmlns:tns="http://www.holidaywebservice.com/HolidayService_v2/">
       <tns:GetHolidaysAvailableResult>
        <tns:HolidayCode><!-- between zero and unlimited repetitions of this element -->
         <tns:Code>string</tns:Code>
         <tns:Description>string</tns:Description>
        </tns:HolidayCode>
       </tns:GetHolidaysAvailableResult>
      </tns:GetHolidaysAvailableResponse>
  GetHolidayDate_tns:
    xml:
      namespace: 'http://www.holidaywebservice.com/HolidayService_v2/'
      prefix: tns
    type: object
    properties:
      countryCode:
        $ref: '#/definitions/Country_tns'
      holidayCode:
        xml:
          namespace: 'http://www.holidaywebservice.com/HolidayService_v2/'
          prefix: tns
        type: string
      year:
        xml:
          namespace: 'http://www.holidaywebservice.com/HolidayService_v2/'
          prefix: tns
        type: integer
        format: int32
    required:
      - countryCode
      - year
    example: >-

      <tns:GetHolidayDate
      xmlns:tns="http://www.holidaywebservice.com/HolidayService_v2/">
       <tns:countryCode><!-- mandatory -->string</tns:countryCode>
       <tns:holidayCode>string</tns:holidayCode>
       <tns:year><!-- mandatory -->3</tns:year>
      </tns:GetHolidayDate>
  GetHolidayDateResponse_tns:
    xml:
      namespace: 'http://www.holidaywebservice.com/HolidayService_v2/'
      prefix: tns
    type: object
    properties:
      GetHolidayDateResult:
        xml:
          namespace: 'http://www.holidaywebservice.com/HolidayService_v2/'
          prefix: tns
        type: string
        format: date-time
    required:
      - GetHolidayDateResult
    example: >-

      <tns:GetHolidayDateResponse
      xmlns:tns="http://www.holidaywebservice.com/HolidayService_v2/">
       <tns:GetHolidayDateResult><!-- mandatory -->2016-04-18T14:07:37</tns:GetHolidayDateResult>
      </tns:GetHolidayDateResponse>
  GetHolidaysForDateRange_tns:
    xml:
      namespace: 'http://www.holidaywebservice.com/HolidayService_v2/'
      prefix: tns
    type: object
    properties:
      countryCode:
        $ref: '#/definitions/Country_tns'
      startDate:
        xml:
          namespace: 'http://www.holidaywebservice.com/HolidayService_v2/'
          prefix: tns
        type: string
        format: date-time
      endDate:
        xml:
          namespace: 'http://www.holidaywebservice.com/HolidayService_v2/'
          prefix: tns
        type: string
        format: date-time
    required:
      - countryCode
      - startDate
      - endDate
    example: >-

      <tns:GetHolidaysForDateRange
      xmlns:tns="http://www.holidaywebservice.com/HolidayService_v2/">
       <tns:countryCode><!-- mandatory -->string</tns:countryCode>
       <tns:startDate><!-- mandatory -->2016-04-18T14:07:37</tns:startDate>
       <tns:endDate><!-- mandatory -->2016-04-18T14:07:37</tns:endDate>
      </tns:GetHolidaysForDateRange>
  GetHolidaysForDateRangeResponse_tns:
    xml:
      namespace: 'http://www.holidaywebservice.com/HolidayService_v2/'
      prefix: tns
    type: object
    properties:
      GetHolidaysForDateRangeResult:
        $ref: '#/definitions/ArrayOfHoliday_tns'
    example: >-

      <tns:GetHolidaysForDateRangeResponse
      xmlns:tns="http://www.holidaywebservice.com/HolidayService_v2/">
       <tns:GetHolidaysForDateRangeResult>
        <tns:Holiday><!-- between zero and unlimited repetitions of this element -->
         <tns:Country><!-- mandatory -->string</tns:Country>
         <tns:HolidayCode>string</tns:HolidayCode>
         <tns:Descriptor>string</tns:Descriptor>
         <tns:HolidayType><!-- mandatory -->string</tns:HolidayType>
         <tns:DateType><!-- mandatory -->string</tns:DateType>
         <tns:BankHoliday><!-- mandatory -->string</tns:BankHoliday>
         <tns:Date><!-- mandatory -->2016-04-18T14:07:37</tns:Date>
         <tns:RelatedHolidayCode>string</tns:RelatedHolidayCode>
         <tns:ApplicableRegions>
          <tns:RegionCode><!-- between zero and unlimited repetitions of this element -->
           <tns:Code>string</tns:Code>
           <tns:Description>string</tns:Description>
          </tns:RegionCode>
         </tns:ApplicableRegions>
        </tns:Holiday>
       </tns:GetHolidaysForDateRangeResult>
      </tns:GetHolidaysForDateRangeResponse>
  GetHolidaysForYear_tns:
    xml:
      namespace: 'http://www.holidaywebservice.com/HolidayService_v2/'
      prefix: tns
    type: object
    properties:
      countryCode:
        $ref: '#/definitions/Country_tns'
      year:
        xml:
          namespace: 'http://www.holidaywebservice.com/HolidayService_v2/'
          prefix: tns
        type: integer
        format: int32
    required:
      - countryCode
      - year
    example: >-

      <tns:GetHolidaysForYear
      xmlns:tns="http://www.holidaywebservice.com/HolidayService_v2/">
       <tns:countryCode><!-- mandatory -->string</tns:countryCode>
       <tns:year><!-- mandatory -->3</tns:year>
      </tns:GetHolidaysForYear>
  GetHolidaysForYearResponse_tns:
    xml:
      namespace: 'http://www.holidaywebservice.com/HolidayService_v2/'
      prefix: tns
    type: object
    properties:
      GetHolidaysForYearResult:
        $ref: '#/definitions/ArrayOfHoliday_tns'
    example: >-

      <tns:GetHolidaysForYearResponse
      xmlns:tns="http://www.holidaywebservice.com/HolidayService_v2/">
       <tns:GetHolidaysForYearResult>
        <tns:Holiday><!-- between zero and unlimited repetitions of this element -->
         <tns:Country><!-- mandatory -->string</tns:Country>
         <tns:HolidayCode>string</tns:HolidayCode>
         <tns:Descriptor>string</tns:Descriptor>
         <tns:HolidayType><!-- mandatory -->string</tns:HolidayType>
         <tns:DateType><!-- mandatory -->string</tns:DateType>
         <tns:BankHoliday><!-- mandatory -->string</tns:BankHoliday>
         <tns:Date><!-- mandatory -->2016-04-18T14:07:37</tns:Date>
         <tns:RelatedHolidayCode>string</tns:RelatedHolidayCode>
         <tns:ApplicableRegions>
          <tns:RegionCode><!-- between zero and unlimited repetitions of this element -->
           <tns:Code>string</tns:Code>
           <tns:Description>string</tns:Description>
          </tns:RegionCode>
         </tns:ApplicableRegions>
        </tns:Holiday>
       </tns:GetHolidaysForYearResult>
      </tns:GetHolidaysForYearResponse>
  GetHolidaysForMonth_tns:
    xml:
      namespace: 'http://www.holidaywebservice.com/HolidayService_v2/'
      prefix: tns
    type: object
    properties:
      countryCode:
        $ref: '#/definitions/Country_tns'
      year:
        xml:
          namespace: 'http://www.holidaywebservice.com/HolidayService_v2/'
          prefix: tns
        type: integer
        format: int32
      month:
        xml:
          namespace: 'http://www.holidaywebservice.com/HolidayService_v2/'
          prefix: tns
        type: integer
        format: int32
    required:
      - countryCode
      - year
      - month
    example: >-

      <tns:GetHolidaysForMonth
      xmlns:tns="http://www.holidaywebservice.com/HolidayService_v2/">
       <tns:countryCode><!-- mandatory -->string</tns:countryCode>
       <tns:year><!-- mandatory -->3</tns:year>
       <tns:month><!-- mandatory -->3</tns:month>
      </tns:GetHolidaysForMonth>
  GetHolidaysForMonthResponse_tns:
    xml:
      namespace: 'http://www.holidaywebservice.com/HolidayService_v2/'
      prefix: tns
    type: object
    properties:
      GetHolidaysForMonthResult:
        $ref: '#/definitions/ArrayOfHoliday_tns'
    example: >-

      <tns:GetHolidaysForMonthResponse
      xmlns:tns="http://www.holidaywebservice.com/HolidayService_v2/">
       <tns:GetHolidaysForMonthResult>
        <tns:Holiday><!-- between zero and unlimited repetitions of this element -->
         <tns:Country><!-- mandatory -->string</tns:Country>
         <tns:HolidayCode>string</tns:HolidayCode>
         <tns:Descriptor>string</tns:Descriptor>
         <tns:HolidayType><!-- mandatory -->string</tns:HolidayType>
         <tns:DateType><!-- mandatory -->string</tns:DateType>
         <tns:BankHoliday><!-- mandatory -->string</tns:BankHoliday>
         <tns:Date><!-- mandatory -->2016-04-18T14:07:37</tns:Date>
         <tns:RelatedHolidayCode>string</tns:RelatedHolidayCode>
         <tns:ApplicableRegions>
          <tns:RegionCode><!-- between zero and unlimited repetitions of this element -->
           <tns:Code>string</tns:Code>
           <tns:Description>string</tns:Description>
          </tns:RegionCode>
         </tns:ApplicableRegions>
        </tns:Holiday>
       </tns:GetHolidaysForMonthResult>
      </tns:GetHolidaysForMonthResponse>
...
  Country_tns:
    xml:
      namespace: 'http://www.holidaywebservice.com/HolidayService_v2/'
      prefix: tns
    type: string
    enum:
      - Canada
      - GreatBritain
      - IrelandNorthern
      - IrelandRepublicOf
      - Scotland
      - UnitedStates
    x-xsi-type: Country
    x-xsi-type-xml:
      namespace: 'http://www.holidaywebservice.com/HolidayService_v2/'
      prefix: tns
    x-xsi-type-uniquename: Country_tns
  ArrayOfHolidayCode_tns:
    xml:
      namespace: 'http://www.holidaywebservice.com/HolidayService_v2/'
      prefix: tns
    type: object
    properties:
      HolidayCode:
        type: array
        items:
          $ref: '#/definitions/HolidayCode_tns'
        x-nullable: true
    x-xsi-type: ArrayOfHolidayCode
    x-xsi-type-xml:
      namespace: 'http://www.holidaywebservice.com/HolidayService_v2/'
      prefix: tns
    x-xsi-type-uniquename: ArrayOfHolidayCode_tns
...